refactor: tighten setup probe ownership
